feat: Refactor vendor context and remove GSTIN field\n\n- Migrated vendor-related modules and tests from the old context to a new context.\n- Renamed VendorContext to VendorService.\n- Renamed VendorContextProvider to VendorServiceProvider.\n- Created VendorRepository to encapsulate vendor-related functions.\n- Updated references to vendor modules across the codebase (e.g., VendorService, VendorServiceProvider, VendorRepository, VendorContext, VendorContextProvider).\n- Moved vendor-related tests from tests/contexts to tests/services.\n- Removed the GSTIN field from the schema and all associated tests and factories.\n- Fixed various test failures due to module relocation and schema changes, including updating test assertions and ensuring correct parameter passing.
