+++ config/config.exs
@@ -107,13 +107,9 @@ config :ex_audit,
     WraftDoc.Pipelines.Stages.Stage,
     WraftDoc.Themes.Theme,
     WraftDoc.Enterprise.ApprovalSystem,
-    WraftDoc.Enterprise.ApprovalSystem,
-    WraftDoc.Enterprise.Flow,
     WraftDoc.Enterprise.Flow,
     WraftDoc.Enterprise.Flow.State,
-    WraftDoc.Enterprise.Flow.State,
-    WraftDoc.Enterprise.Vendor,
-    WraftDoc.Enterprise.Vendor,
+    WraftDoc.Vendors.Vendor,
     WraftDoc.Forms.Form,
     WraftDoc.Layouts.Layout,
     WraftDoc.Layouts.LayoutAsset,
